Introducing Anchor (#37)
Adds `RopeAnchor` class, with different behaviors.
- ANCHORED: Rope stays put
- TOWING: Rope pulls on parent of anchor if parent is a RigidBody3D
- GUIDED: Rope slides through anchor point
- SLIDING: Guided & Towing. Useful for pulleys

Adds `CoiledAnchor` class. Behavior is applied at first position.

Removes:
- PullRope
- RopeAnchorsBase
- RopeAttachmentsBase
- CoiledRope

+++ src/wave_sampler.cpp
@@ -478,8 +478,8 @@ Color WaveSampler::_sample_image(const Ref<Image> &p_image, const Vector2 &uv, c
 	}
 
 	// Handle wrapping (fract in GLSL)
-	float u = Math::fmod(uv.x, 1.0f);
-	float v = Math::fmod(uv.y, 1.0f);
+	float u = Math::fmod((float)uv.x, 1.0f);
+	float v = Math::fmod((float)uv.y, 1.0f);
 	if (u < 0.0f)
 		u += 1.0f;
 	if (v < 0.0f)
